backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
langchain
langchain-google-genai
langsmith
langgraph
google-generativeai
httpx
faiss-cpu
sqlalchemy
llama-index
llama-index-vector-stores-faiss
llama-index-embeddings-google
pandas
openpyxl

